+++ packages/components/src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './GlobalStyle'
+export * from './useControlWarn'
 export * from './useToggle'